Use resource key as name in permissions code
The code relied on the `Name` property being accessible for every resource.
This is generally true, but because these property structs are embedded as
pointer, they can be nil. This is also why the tests had to initialize the
embedded struct to pass. This changes the approach to use the keys from the
resource map instead, so that we no longer rely on the non-nil embedded struct.

Note: we should evaluate whether we should turn these into values instead of
pointers. I don't recall if we get value from them being pointers.