Support passing job parameters to bundle run
This change adds support for job parameters. If job parameters are specified
for a job that doesn't define job parameters it returns an error. Conversely,
if task parameters are specified for a job that defines job parameters, it also
returns an error.

This changes moves the options structs and their functions to separate files
and backfills test coverage for them.

Job parameters can now be specified with `--param foo=bar,bar=qux`.