Prompt for account profile only for account-level command execution instead of during `databricks labs install` flow
There's a lot of end-user friction for projects, that require account-level commands. This is mainly related to the fact, that, as of January 2024, workspace administrators do not necessarily have access to call account-level APIs. There are ongoing discussions on how to best implement this on a platform level.

Current temporary workaround is creating dummy ~/.databrickscfg profile with `account_id` field, though it doesn't really remove the end-user friction, hence we don't require an account profile during installation (anymore) and just prompt for it, when context  does require it. This also means that we always prompt for account-level commands, unless users specify a `--profile` flag.