Retain partially valid structs in `convert.Normalize`
Before this change, any error in a subtree would cause the entire subtree to be
dropped from the output.

This is not ideal when debugging, so instead we drop only the values that
cannot be normalized. Note that this doesn't change behavior if the caller is
properly checking the returned diagnostics for errors.