Add option to include fields present in the type but not in the value
This feature supports variable lookups in a `dyn.Value` that are present in the
type but haven't been initialized with a value.

For example: `${bundle.git.origin_url}` is present in the `dyn.Value` only if
it was assigned a value. If it wasn't assigned a value it should resolve to the
empty string. This normalization option, when set, ensures that all fields that
are represented in the specified type are present in the return value.