Use `any` as type for data sources and resources in `tf/schema`
We plan to use the any-equivalent of a `dyn.Value` such that we can use
variable references for non-string fields (e.g. `${databricks_job.some_job.id}`
where an integer is expected), as well as properly emit zero values for
primitive types (e.g. 0 for integers or false for booleans).

This change is in preparation for the above.