Inline logic to set a value in `dyn.SetByPath`
This removes the need for the `allowMissingKeyInMap` option to the private
`visit` function and ensures that the body of the visit function doesn't add or
remove values of the configuration it traverses.

This in turn prepares for visiting a path pattern that yields more than one
callback, which doesn't match well with the now removed option.