Add `dyn.MapByPattern` to map a function to values with matching paths
The new `dyn.Pattern` type represents a path pattern that can match one or more
paths in a configuration tree. Every `dyn.Path` can be converted to a
`dyn.Pattern` that matches only a single path.

To accommodate this change, the visit function needed to be modified to take a
`dyn.Pattern` suffix. Every component in the pattern implements an interface to
work with the visit function. This function can recurse on the visit function
for one or more elements of the value being visited. For patterns derived from
a `dyn.Path`, it will work as it did before and select the matching element.
For the new pattern components (e.g. `dyn.AnyKey` or `dyn.AnyIndex`), it
recurses on all the elements in the container.