Define `dyn.Mapping` to represent maps
Before this change maps were stored as a regular Go map with string keys. This
didn't let us capture metadata (location information) for map keys.

To address this, this change replaces use of the regular Go map with a
dedicated type for a dynamic map. This type stores the `dyn.Value` for both the
key and the value. It uses a map to still allow O(1) lookups and redirects
those into a slice.