Convert between integer and float in normalization
We currently issue a warning if an integer is used where a floating point
number is expected. But if they are convertible, we should convert and not
issue a warning. This change fixes normalization if they are convertible
between each other. We still produce a warning if the type conversion leads to
a loss in precision.