Don't merge-in remote resources during depolyments
`check_running_resources` now pulls the remote state without modifying the bundle state,
similar to how it was doing before. This avoids a problem when we fail to compute deployment
metadata for a deleted job (which we shouldn't do in the first place)

`deploy_then_remove_resources_test` now also deploys and deletes a job (in addition to a pipeline),
which catches the error that this PR fixes.