Don't hide commands of services that are already hidden
Currently, the help output of services in preview doesn't show any of their
commands because the commands themselves are hidden as well.

This change updates that behavior to not hide commands in preview if the
service itself is also in preview. This makes the help output of services in
preview actually usable.